Refactoring(server-info): removed separate cli class, its Server's args
